// src/models/client.ts
export class Client{
    id: string;
    name: string;
    lastName: string;
    phoneNumber: number;
    location: Location;
    photo: String;

    // constructor(name: string, lastName: string, phoneNumber: number){
    //     this.name = name;
    //     this.lastName = lastName;
    //     this.phoneNumber = phoneNumber;        
    // }

    constructor(){
        
    }
}